[test] gathering test related to the sign instructions
