--- tests/integration/nano/test_sign_micheline_without_hash.py
@@ -1,42 +1,0 @@
-#!/usr/bin/env python3
-# Copyright 2023 Functori <contact@functori.com>
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-# http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Check signing micheline wihout getting hash"""
-
-from pathlib import Path
-
-from utils.app import Screen, TezosAppScreen, DEFAULT_ACCOUNT
-from utils.message import MichelineExpr
-
-def test_sign_micheline_without_hash(app: TezosAppScreen):
-    """Check signing micheline wihout getting hash"""
-    test_name = Path(__file__).stem
-
-    app.assert_screen(Screen.HOME)
-
-    message = MichelineExpr([{'string': 'CACA'}, {'string': 'POPO'}, {'string': 'BOUDIN'}])
-
-    data = app.sign(DEFAULT_ACCOUNT,
-                    message,
-                    with_hash=False,
-                    path=test_name)
-
-    app.checker.check_signature(
-        account=DEFAULT_ACCOUNT,
-        message=message,
-        with_hash=False,
-        data=data)
-
-    app.quit()
